+++ src/components/commons/SocialLoginBox/index.tsx
@@ -1,3 +1,4 @@
+import classNames from "classnames";
 import { IconGoogle, IconKakaotalk } from "@utils/icon";
 
 type SocialLoginBoxProps = {
@@ -13,14 +14,16 @@ type SocialLoginBoxProps = {
  */
 
 export default function SocialLoginBox({ type, onGoogleClick, onKakaoClick }: SocialLoginBoxProps) {
+  const classnames = classNames("w-full md:w-460", type === "register" && "mt-24 md:mt-48");
+
   return (
-    <div className="w-full md:w-460">
+    <div className={classnames}>
       <div className="mb-16 flex items-center gap-24">
         <div className="h-1 flex-1 border border-solid border-border-primary" />
         <span className="text-lg font-normal text-text-inverse">OR</span>
         <div className="h-1 flex-1 border border-solid border-border-primary" />
       </div>
-      <div className="flex justify-between">
+      <div className="flex items-center justify-between">
         <span className="text-lg font-medium text-text-inverse">{`간편 ${type === "login" ? "로그인" : "회원가입"}하기`}</span>
         <div className="flex gap-16">
           <button
